Split components and utility functions into specific files
