+++ dragon-boat/App.scss
@@ -18,138 +18,4 @@
 		align-items: center;
 		gap: 12px;
 	}
-
-	.toggles {
-		width: 300px;
-		display: flex;
-		flex-direction: column;
-		gap: 8px;
-
-		section {
-			display: flex;
-			flex-direction: column;
-		}
-	}
-
-	.boat-container {
-		text-align: center;
-
-		.boat {
-			display: flex;
-			flex-direction: row;
-			gap: 4px;
-
-			.paddlers {
-				display: flex;
-				flex-direction: column;
-				gap: 4px;
-			}
-
-			.center {
-				display: flex;
-				flex-direction: column;
-				justify-content: space-between;
-			}
-
-			.boat-stats {
-				display: flex;
-				flex-direction: column;
-				gap: 4px;
-			}
-		}
-	}
-
-	.roster-container {
-		display: flex;
-		flex-direction: column;
-		align-items: center;
-		gap: 4px;
-
-		.button-group {
-			display: flex;
-			flex-direction: row;
-			gap: 12px;
-		}
-
-		.roster {
-			position: relative;
-
-			display: grid;
-			grid-template-columns: repeat(3, 1fr);
-			gap: 4px;
-		}
-	}
-}
-
-.paddler-card {
-	width: 92px;
-	height: 40px;
-	border-radius: 4px;
-	padding: 6px 8px;
-
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-
-	opacity: 1;
-	touch-action: none;
-
-	.name {
-		overflow: hidden;
-		white-space: nowrap;
-		text-overflow: ellipsis;
-	}
-
-	.info {
-		display: flex;
-		flex-direction: column;
-		align-items: end;
-		margin-left: 4px;
-
-		.side {
-			font-weight: bold;
-		}
-
-		.left {
-			color: var(--red);
-		}
-
-		.right {
-			color: var(--cyan);
-		}
-
-		.both {
-			color: var(--purple);
-		}
-	}
-
-	.empty-slot {
-		width: 100%;
-		text-align: center;
-	}
-}
-
-.details {
-	background-color: var(--light);
-	color: var(--darkest);
-	cursor: pointer;
-}
-
-.empty {
-	background-color: rgba(0, 0, 0, 0.3);
-	color: var(--light);
-}
-
-.dragging {
-	opacity: 0.8;
-}
-
-.dragged-over {
-	opacity: 0.5;
-	background-color: var(--cyan);
-	color: var(--darkest);
-}
-
-.return {
-	background-color: var(--red);
 }
