+++ dragon-boat/components/Boat/Boat.scss
@@ -1,0 +1,29 @@
+@import url(../../colors.css);
+
+.boat-container {
+	text-align: center;
+
+	.boat {
+		display: flex;
+		flex-direction: row;
+		gap: 4px;
+
+		.paddlers {
+			display: flex;
+			flex-direction: column;
+			gap: 4px;
+		}
+
+		.center {
+			display: flex;
+			flex-direction: column;
+			justify-content: space-between;
+		}
+
+		.boat-stats {
+			display: flex;
+			flex-direction: column;
+			gap: 4px;
+		}
+	}
+}
