+++ dragon-boat/components/PaddlerCard/PaddlerCard.scss
@@ -1,0 +1,74 @@
+@import url(../../colors.css);
+
+.paddler-card {
+	width: 92px;
+	height: 40px;
+	border-radius: 4px;
+	padding: 6px 8px;
+
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
+
+	opacity: 1;
+	touch-action: none;
+
+	.name {
+		overflow: hidden;
+		white-space: nowrap;
+		text-overflow: ellipsis;
+	}
+
+	.info {
+		display: flex;
+		flex-direction: column;
+		align-items: end;
+		margin-left: 4px;
+
+		.side {
+			font-weight: bold;
+		}
+
+		.left {
+			color: var(--red);
+		}
+
+		.right {
+			color: var(--cyan);
+		}
+
+		.both {
+			color: var(--purple);
+		}
+	}
+
+	.empty-slot {
+		width: 100%;
+		text-align: center;
+	}
+}
+
+.details {
+	background-color: var(--light);
+	color: var(--darkest);
+	cursor: pointer;
+}
+
+.empty {
+	background-color: rgba(0, 0, 0, 0.3);
+	color: var(--light);
+}
+
+.dragging {
+	opacity: 0.8;
+}
+
+.dragged-over {
+	opacity: 0.5;
+	background-color: var(--cyan);
+	color: var(--darkest);
+}
+
+.return {
+	background-color: var(--red);
+}
