+++ dragon-boat/components/Roster/Roster.scss
@@ -1,0 +1,22 @@
+@import url(../../colors.css);
+
+.roster-container {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	gap: 4px;
+
+	.button-group {
+		display: flex;
+		flex-direction: row;
+		gap: 12px;
+	}
+
+	.roster {
+		position: relative;
+
+		display: grid;
+		grid-template-columns: repeat(3, 1fr);
+		gap: 4px;
+	}
+}
